Creo que esto soluciona el problema que mando Robert

+++ trunk/booktube/src/main/java/com/booktube/pages/BooksByTag.java
@@ -94,7 +94,7 @@ public class BooksByTag extends BasePage {
 					public void onClick() {
 
 						Book book = (Book) getModelObject();
-						Integer bookId = book.getId();
+						Long bookId = book.getId();
 
 						bookService.deleteBook(book);
 						System.out.println("Book " + bookId + " deleted.");
